% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asy_vzt.R
\name{asy_vzt}
\alias{asy_vzt}
\title{Asymptotischer Vorzeichen-Rangtest}
\usage{
asy_vzt(n, alternative = c("less", "two.sided", "greater"))
}
\arguments{
\item{n}{Stichprobengroesse}

\item{alternative}{Richtung der Hypothese}
}
\value{
Kritischer Wert
}
\description{
Einfache Funktion zur Berechnung des asymptotischen Vorzeichen-Rangtests
}
