% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asy_wrt.R
\name{asy_wrst}
\alias{asy_wrst}
\title{Asymptotischer Wilcoxon-Rangsummentest}
\usage{
asy_wrst(n, ni, ri, t = 1)
}
\arguments{
\item{n}{Stichprobengrösse}

\item{ni}{n1 oder n2}

\item{ri}{RS1 oder RS2}

\item{t}{Tiekorrektur}
}
\value{
Erwartungswert, Varianz und empirischer z-Wert
}
\description{
Asymptotischer Wilcoxon-Rangsummentest
}
\examples{
# Fall ohne Ties
asy_wrst(n = 7, ni = 3, ri = 11)

# Fall mit Ties
asy_wrst(n = 7, ni = 3, ri = 11, t = 0.9107)
}
