% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asy_wrt.R
\name{asy_wrt}
\alias{asy_wrt}
\title{Asymptotischer Wilcoxon-Rangsummentest}
\usage{
asy_wrt(n, ni, ri, t)
}
\arguments{
\item{n}{Stichprobengrösse}

\item{ni}{n1 oder n2}

\item{ri}{RS1 oder RS2}

\item{t}{Tiekorrektur}
}
\value{
Erwartungswert, Varianz und empirischer z-Wert
}
\description{
Asymptotischer Wilcoxon-Rangsummentest
}
