% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomTable.R
\name{binomTable}
\alias{binomTable}
\title{Binomialverteilungstabelle}
\usage{
binomTable(n, probability)
}
\arguments{
\item{n}{Stichprobengroesse}

\item{probability}{Wahrscheinlichkeit des Ereignisses}
}
\value{
Binomialverteilungstabelle
}
\description{
Erstellt eine Binomialverteilungstabelle basierend auf der Stichprobengrösse und der Wahrscheinlichkeit Pi
}
