% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_plot.R
\name{chi.plot}
\alias{chi.plot}
\title{Chi-Quadrat-Verteilung Plot}
\usage{
chi.plot(chi, df, x)
}
\arguments{
\item{chi}{Empirisches Chi-Quadrat}

\item{df}{Freiheitsgrade}

\item{x}{Definiert Breite des Plots. Je höher df und chi, desto höher x}
}
\description{
Plottet die Chi-Quadrat Verteilung
}
