% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cochrans_effsize.R
\name{cochrans_effsize}
\alias{cochrans_effsize}
\title{Effect Sizes for Cochrans Q}
\usage{
cochrans_effsize(
  data,
  formula = NULL,
  dep = NULL,
  samp = NULL,
  block = NULL,
  effsize = c("R", "eta2")
)
}
\arguments{
\item{data}{Dataset in long format}

\item{formula}{Formula in the format "DV ~ grouping variable | block"}

\item{dep}{Dependent variable}

\item{samp}{Grouping variable}

\item{block}{Block (person variable / ID)}

\item{effsize}{Effect size of interest, i.e. eta2 or R}
}
\value{
Returns the effect size of interest (values between -1 and +1)
}
\description{
This function calculates the widely used effect size eta2, but also the less prevalent
but in many ways more adequate effect size R. See Berry et al. (2007) for a discussion:
https://journals.sagepub.com/doi/10.2466/pms.104.4.1236-1242
}
\examples{
# Dataset from Berry et al. (2007)
berry <- data.frame(subject = rep(1:6, each = 8),
                    treatment = rep(1:8, times = 6),
                    correct = c(0,1,1,1,0,0,1,0,
                                1,1,1,0,0,1,1,1,
                                0,1,0,1,1,0,1,1,
                                1,1,1,1,0,1,1,1,
                                0,1,1,0,0,0,1,1,
                                1,1,1,1,0,1,1,0))

# Get the effect size R by specifing dep, samp and block 
cochrans_effsize(data = berry, 
                 dep = "correct", 
                 samp = "treatment", 
                 block = "subject", 
                 effsize = "R")

# Get the effect size R by providing a formula
formula = as.formula(correct ~ treatment | subject)
cochrans_effsize(data = berry, 
                 formula = formula, 
                 effsize = "R")

# Get the effect size eta2 by specifing dep, samp and block 
cochrans_effsize(data = berry, 
                 dep = "correct", 
                 samp = "treatment", 
                 block = "subject", 
                 effsize = "eta2")

# Get the effect size eta2 by providing a formula
cochrans_effsize(data = berry, 
                 formula = formula, 
                 effsize = "eta2")
}
