% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erwartungswerte.R
\name{erwartungswerte}
\alias{erwartungswerte}
\title{Erwartungswerte}
\usage{
erwartungswerte(data)
}
\arguments{
\item{data}{Matrix mit den beobachteten Werten}
}
\value{
Matrix mit den erwarteten Werten
}
\description{
Berechnet die Erwartungswerte im Chi-Quadrat-Test
}
