% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_2x2_pw.R
\name{exact_2xk_pw}
\alias{exact_2xk_pw}
\title{Punktwahrscheinlichkeit exakter 2xk-Tests}
\usage{
exact_2xk_pw(data)
}
\arguments{
\item{data}{Matrix mit den beobachteten Werten}
}
\value{
Matrix mit den erwarteten Werten
}
\description{
Berechnet die Punktwahrscheinlichkeit des exakten 2xk-Tests
}
