% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperTable.R
\name{hyperTable}
\alias{hyperTable}
\title{hyperTable}
\usage{
hyperTable(n1p, np1, n)
}
\arguments{
\item{n1p}{Grösse der Stichprobe 1}

\item{np1}{Anzahl Personen mit Merkmal 1}

\item{n}{Gesamte Stichprobengrösse}
}
\value{
Hyperparametrische Verteilungstabelle
}
\description{
Berechnet eine Tabelle der hypergeometrisch verteilten Wahrscheinlichkeiten basierend auf n1p, np1 und n
}
\examples{
hyperTable(6, 6, 13)
}
