% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw_slnh.R
\name{kw_slnh}
\alias{kw_slnh}
\title{Kruskal-Wallis-Test: Spezielle Lage-Nullhypothesen}
\usage{
kw_slnh(data, ...)
}
\arguments{
\item{data}{Matrix mit den beobachteten Werten, wobei jede Reihe eine Gruppe ist.}

\item{...}{Platzhalter für Kontraste. Anzahl Kontraste muss = Stichprobengroesse sein.}
}
\value{
Teststatistik L, Tiekorrektur, Erwartungswert, empirischer z-Wert und kritische L-Werte
}
\description{
Experimentelle Funktion zur Berechnung der speziellen Lage-Nullhypothesen des Kruskal-Wallis-Tests basierend auf Kontrasten.
}
\examples{
data <- matrix(c(5,8,7, 4,4,NA, 2,5,NA, 6,9,6), nrow = 4, ncol = 3, byrow = TRUE)

# Gruppengleichheitshypothese
kw_slnh(data, 1, -1, 0, 0)

# Lineare Interaktion
kw_slnh(data, 1, -1, -1, 1)
}
