% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peachi_test.R
\name{peachi_test}
\alias{peachi_test}
\title{Pearsons Chi-Quadrat Test}
\usage{
peachi_test(data)
}
\arguments{
\item{data}{Matrix mit den beobachteten Werten}
}
\value{
n_p2, n_p•, Chi-Quadrat und df, kritischer Bereich, p-Wert und Interpretation
}
\description{
Berechnet Pearsons Chi-Quadrat Test fuer unbestimmte Anzahl Kategorien
}
