% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrst.R
\name{wrsTable}
\alias{wrsTable}
\title{Wilcoxon Rangsummen-Test}
\usage{
wrsTable(s1, s2)
}
\arguments{
\item{s1}{Werte der Stichprobe 1}

\item{s2}{Werte der Stichprobe 2}
}
\value{
Originalrang-Kombinationen und Verteilungstabelle
}
\description{
Berechnet die Originalrang-Kombinationen und die exakte Verteilung im Wilcoxon Rangsummen-Test.
Die Funktion dient primär zur Unterstützung/Kontrolle bei der Berechnung von Hand.
}
\examples{
s1 <- c(1,4,9)
s2 <- c(8,3,6,7)
wrsTable(s1, s2)
}
