% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw_slnh.R
\name{kw_slnh}
\alias{kw_slnh}
\title{Kruskal-Wallis-Test: Spezielle Lag-Nullhypothesen}
\usage{
kw_slnh(data, ...)
}
\arguments{
\item{data}{Matrix mit den beobachteten Werten}

\item{...}{Platzhalter für Kontraste. Anzahl Kontraste muss = Stichprobengroesse sein.}
}
\value{
Teststatistik L, Tiekorrektur, Erwartungswert, empirischer z-Wert und kritische L-Werte
}
\description{
Experimentelle Funktion zur Berechnung der speziellen Lage-Nullhypothesen des Kruskal-Wallis-Tests basierend auf Kontrasten.
}
