% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrst.R
\name{wrst}
\alias{wrst}
\title{Wilcoxon Rangsummen-Test}
\usage{
wrst(n1, n2)
}
\arguments{
\item{n1}{n1}

\item{n2}{n2}
}
\value{
Originalrang-Kombinationen und Verteilungstabelle
}
\description{
Berechnet die exakte Verteilung im Wilcoxon Rangsummen-Test
}
