% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrst.R
\name{wrst}
\alias{wrst}
\title{Wilcoxon Rangsummen-Test}
\usage{
wrst(s1, s2)
}
\arguments{
\item{s1}{Werte der Stichprobe 1}

\item{s2}{Werte der Stichprobe 2}
}
\value{
Originalrang-Kombinationen und Verteilungstabelle
}
\description{
Berechnet die Originalrang-Kombinationen und die exakte Verteilung im Wilcoxon Rangsummen-Test.
}
